working on texture load
