+++ core/src/lib.rs
@@ -8,3 +8,5 @@ pub mod factory;
 ///Type alias for nannou color type
 pub type Mrgb = Rgb<Srgb, u8>;
 
+
+
