+++ core/src/soots.rs
@@ -1,1 +1,2 @@
+pub mod sootsprite;
 pub mod arthropods;
